lambda reports/requirements.txt
redshift_connector
python-dotenv